+++ tests/test_sha384.py
@@ -3,8 +3,8 @@ import unittest
 from hashbase import SHA384
 
 
-class TestSHA1Strings(unittest.TestCase):
-    def test_sha1_string_input(self):
+class TestSHA384Strings(unittest.TestCase):
+    def test_sha384_string_input(self):
         self.assertEqual(
             SHA384().generate_hash(""),
             "38b060a751ac96384cd9327eb1b1e36a21fdb71114be07434c0cc7bf63f6e1da274edebfe76f65fbd51ad2f14898b95b",